feat: added a command to print all the stars and the ASCII art
* Added a `stars` command to print the ASCII art and the star progress for a given year.

Closes #27

+++ src/domain/ports/get_leaderboard.rs
@@ -3,5 +3,5 @@ use super::super::leaderboard::Leaderboard;
 use super::errors::AocClientError;
 
 pub trait GetLeaderboard {
-    fn get_leaderboard(&self, year: u16) -> Result<Leaderboard, AocClientError>;
+    fn get_leaderboard(&self, year: i32) -> Result<Leaderboard, AocClientError>;
 }
